Fixed bug that broken resource compilation.
When the GUI CMakeList was modified for support for
Qt5 the RCC resources were no longer being compiled
into the binaries due to a naming change. This broke
icons within the application and was most visible in
the data manager. Correcting the name fixes this issue
and icons now properly display again.