Add Teem library for nrrd support.
This change adds the Teem nrrd library back to the build tree.
This library is used to load and handle nrrd files without the need
to have 3rd party software installed on the user machine. The cmake
flag for USE_ITK still defaults to ON. Users will have to turn this
flag off to use the Teem library instead.

There is a warning output when generating the make files using teem.