Updates travis-ci yaml script to also build cleaver's CLI.
For a period of time ITK was a required dependency necessary for
building the cleaver CLI. It enabled the loading of NRRD files.
Now that the Teem library is back in the source tree, the CLI can
be built against the Teem library without requiring ITK. This also
allows us to begin building the Cleaver-CLI as part of continuous
integration for testing.