Fix broken cmake builds for cli regression tests.
There is a directory of CLI regression tests which have been
commented out for a considerable time. The newtest macro has
been updated to include additional shared source code. This
build pattern follows the previous intended macro of cxx_test
and rebuilds the shared code for each new test. This should
be fixed at some point to only compile the shared source once
and link the result against any tests that use it.

Most of the regression tests appear to be broken, with the
exception of the 'cli_strict' test. The rest of the tests will
remain disabled until the cause of the failure is determined
or some other appropriate course of action is taken.

Added whitespacing to cli_strict test to improve readability.