Fixed field type on NRRDTools::saveNRRDFile.
When the ITK version of NRRDTools was introduced, the function
signature was relaxed from a FloatField to an AbstractScalarField.
This could result in unpredictable sampling of a field on export.
Any conversion from an implicit field to a scalar sampling should
happen prior to this method being called to ensure there is no
undesired data loss.