Convert interface order #defines to strictly typed enum class Order.
This change improves the type safety for vertex interface order by
creating an enum class named Order in the cleaver namespace. With
only 4 possible values, the int8_t can be used to save memory.

This increased type safety caught a bug in the stencil logic which
counts the number of edges. The fix is included in this change.