Fixes pimpl ideom that was broken by commit ab76493fa09ee2be75b5e2591010ea7800d2153d
This change fixes the pimpl idiom used in the CleaverMesher. The idiom hides the actual
interface from the front end to prevent accidental or intentional tight coupling of the
implementation details. Commit ab76493fa09ee2be75b5e2591010ea7800d2153d inverted this
idiom, exposing implementation details to the main interface and defeating the purpose
of the pattern.