Add recording flag to command line and mesher.
A debug.dump file is provided as input for recording. If recording
is enabled, any cleaver operations that affect the tets found in
the input dump file are recorded and saved to an operations list.
This list can then be used to replay the events that led to the
poor quality tets, without the larg memory footpring of the rest
of the mesh.

This change only parses the input debug and saves a list of the
parent tets that generated the output tets in the dump.