Automatically add .nrrd extension to exported float fields.
There's a bug in Qt on MacOs where calling to
QFileDialog::getSaveFileName does not open a Qt Dialog but
instead opens the native file open dialog. The native dialog
ignores the extension information provided.

The above error is compounded by the introduction of ITK as
a mechanism for saving nrrd files. It expects the name given
to include the file extension, and throws an exception if one
is not provided.