Add option to disable interfaces when serializing tets to json.
Part of the new debug system is code for serializing tetrahedron
information to json format. It is primarily used in recording
mode, but is also used to print out the initial dump of bad tets.
This initial dump may include tets for which there are not
interfaces. If they are included, an invalid memory reference will
result in a crash. This should now be prevented.