Move dihedral angles menu option to compute submenu.
It looks like at some point the compute menu was removed from the
GUI. This change adds it back and moves the action to compute
dihedral angles into this submenu. This is more intuitive than
listing it under the 'edit' menu, since computing dihedral angles
is not an action that modifies the mesh or visualization.